Don't merge skipped conditions from methods; see #14
The MS instrument sometimes "skips" conditions that are present in the
method file. There aren't any scans for this particular condition in the
raw file (because of that there is no information in the header and the
mzML file). Before we kept these conditions by using
`merge(..., all.y=TRUE)`. The current commit removes such "skipped"
conditions.

See also:
- https://github.com/sgibb/topdown/issues/14#issuecomment-320533032
- https://github.com/sgibb/topdown/issues/14#issuecomment-320535814